+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix `RSpec/DescribedClass`'s error when a `described_class` is part of the namespace. ([@pirj][])
 * Fix `RSpec/ExampleWording` autocorrect of multi-line docstrings. ([@pirj][])
 * Add `RSpec/ContextMethod` cop, to detect method names in `context`. ([@geniou][])
+* Update RuboCop dependency to 0.68.1 with support for children matching node pattern syntax. ([@pirj][])
 
 ## 1.35.0 (2019-08-02)